Renamed the api namespace to os!
The api namespace will be specific for GPU processing APIs. The os namespace includes all classes that are OS specific
