+++ CriaAI-Core/CriaAI-Core.vcxproj
@@ -101,21 +101,22 @@ copy "$(CudaToolkitBinDir)\cudart*.dll" "$(OutDir)"</Command>
     <ClCompile Include="Dependencies\BmpRenderer\src\API\ScreenDrawer.cpp" />
     <ClCompile Include="Dependencies\BmpRenderer\src\Bitmap.cpp" />
     <ClCompile Include="Dependencies\BmpRenderer\src\Renderer.cpp" />
-    <ClCompile Include="src\api\FileSystem.cpp" />
-    <ClCompile Include="src\api\InputLogger.cpp" />
-    <ClCompile Include="src\api\InputUtil.cpp" />
-    <ClCompile Include="src\api\OSContext.cpp" />
-    <ClCompile Include="src\api\ScreenCapturer.cpp" />
-    <ClCompile Include="src\api\Window.cpp" />
-    <ClCompile Include="src\api\win\WinInputLogger.cpp" />
-    <ClCompile Include="src\api\win\WinInputSimulator.cpp" />
-    <ClCompile Include="src\api\win\WinOSContext.cpp" />
-    <ClCompile Include="src\api\win\WinScreenCapturer.cpp" />
-    <ClCompile Include="src\api\InputSimulator.cpp" />
-    <ClCompile Include="src\api\win\WinWindow.cpp" />
+    <ClCompile Include="src\os\FileSystem.cpp" />
+    <ClCompile Include="src\os\InputLogger.cpp" />
+    <ClCompile Include="src\os\InputUtil.cpp" />
+    <ClCompile Include="src\os\OSContext.cpp" />
+    <ClCompile Include="src\os\ScreenCapturer.cpp" />
+    <ClCompile Include="src\os\Window.cpp" />
+    <ClCompile Include="src\os\win\WinInputLogger.cpp" />
+    <ClCompile Include="src\os\win\WinInputSimulator.cpp" />
+    <ClCompile Include="src\os\win\WinOSContext.cpp" />
+    <ClCompile Include="src\os\win\WinScreenCapturer.cpp" />
+    <ClCompile Include="src\os\InputSimulator.cpp" />
+    <ClCompile Include="src\os\win\WinWindow.cpp" />
     <ClCompile Include="src\maths\Matrixf.cpp" />
     <ClCompile Include="src\network\NeuronGroup.cpp" />
     <ClCompile Include="src\network\NeuronNetwork.cpp" />
+    <CudaCompile Include="src\network\ActivationFunctions.cu" />
     <CudaCompile Include="src\util\FloatBitmap.cu">
       <FileType>Document</FileType>
     </CudaCompile>
@@ -133,19 +134,19 @@ copy "$(CudaToolkitBinDir)\cudart*.dll" "$(OutDir)"</Command>
     <ClInclude Include="Dependencies\BmpRenderer\src\Bitmap.hpp" />
     <ClInclude Include="Dependencies\BmpRenderer\src\Common.h" />
     <ClInclude Include="Dependencies\BmpRenderer\src\Renderer.hpp" />
-    <ClInclude Include="src\api\cuda\CudaContext.cuh" />
-    <ClInclude Include="src\api\FileSystem.h" />
-    <ClInclude Include="src\api\InputLogger.h" />
-    <ClInclude Include="src\api\InputSimulator.h" />
-    <ClInclude Include="src\api\InputUtil.h" />
-    <ClInclude Include="src\api\OSContext.h" />
-    <ClInclude Include="src\api\ScreenCapturer.h" />
-    <ClInclude Include="src\api\Window.h" />
-    <ClInclude Include="src\api\win\WinOSContext.h" />
-    <ClInclude Include="src\api\win\WinInputLogger.h" />
-    <ClInclude Include="src\api\win\WinInputSimulator.h" />
-    <ClInclude Include="src\api\win\WinScreenCapturer.h" />
-    <ClInclude Include="src\api\win\WinWindow.h" />
+    <ClInclude Include="src\os\cuda\CudaContext.cuh" />
+    <ClInclude Include="src\os\FileSystem.h" />
+    <ClInclude Include="src\os\InputLogger.h" />
+    <ClInclude Include="src\os\InputSimulator.h" />
+    <ClInclude Include="src\os\InputUtil.h" />
+    <ClInclude Include="src\os\OSContext.h" />
+    <ClInclude Include="src\os\ScreenCapturer.h" />
+    <ClInclude Include="src\os\Window.h" />
+    <ClInclude Include="src\os\win\WinOSContext.h" />
+    <ClInclude Include="src\os\win\WinInputLogger.h" />
+    <ClInclude Include="src\os\win\WinInputSimulator.h" />
+    <ClInclude Include="src\os\win\WinScreenCapturer.h" />
+    <ClInclude Include="src\os\win\WinWindow.h" />
     <ClInclude Include="src\Common.hpp" />
     <ClInclude Include="src\Macros.hpp" />
     <ClInclude Include="src\maths\Maths.h" />
@@ -153,6 +154,7 @@ copy "$(CudaToolkitBinDir)\cudart*.dll" "$(OutDir)"</Command>
     <ClInclude Include="src\maths\Rectangle2D.h" />
     <ClInclude Include="src\maths\Vec2.h" />
     <ClInclude Include="src\maths\Vec3.h" />
+    <ClInclude Include="src\network\ActivationFunctions.h" />
     <ClInclude Include="src\network\NeuronLayer.h" />
     <ClInclude Include="src\network\NetworkUtil.h" />
     <ClInclude Include="src\network\NeuronNetwork.h" />
@@ -167,7 +169,7 @@ copy "$(CudaToolkitBinDir)\cudart*.dll" "$(OutDir)"</Command>
     <ClInclude Include="src\util\Random.h" />
   </ItemGroup>
   <ItemGroup>
-    <CudaCompile Include="src\api\cuda\CudaContext.cu" />
+    <CudaCompile Include="src\os\cuda\CudaContext.cu" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
