+++ CriaAI-Core/src/os/OSContext.h
@@ -1,4 +1,4 @@
-/******************************************************************************
+﻿/******************************************************************************
 * Cria  - The worst artificial intelligence on the market.                    *
 *         <https://github.com/xFrednet/CriaAI>                                *
 *                                                                             *
@@ -32,82 +32,47 @@
 ******************************************************************************/
 #pragma once
 
-#include "NetworkUtil.h"
+#include "../Common.hpp"
 
-#include "../os/InputLogger.h"
-#include "../os/InputSimulator.h"
-
-#ifndef CR_NEURON_OUTPUT_ACTIVATION_LIMIT
-#	define CR_NEURON_OUTPUT_ACTIVATION_LIMIT     crnwdec(0.5)
-#endif
-
-namespace cria_ai { namespace network {
+namespace cria_ai { namespace os {
 	
-	typedef enum CR_NEURON_TYPE_ {
-		CR_NEURON_INPUT_LAYER        = 0x0100,
-		CR_NEURON_DATA_INPUT         = 0x0101,
-		CR_NEURON_RANDNOISE_INPUT    = 0x0102,
-
-		CR_NEURON_HIDDEN_LAYER       = 0x0200,
-		CR_NEURON_NORMAL             = 0x0201,
-		CR_NEURON_INPUT_DELAY        = 0x0203,
-
-		CR_NEURON_OUTPUT_LAYER       = 0x0400,
-		CR_NEURON_KEYBOARD_OUTPUT    = 0x0401,
-		CR_NEURON_MMOVE_OUTPUT       = 0x0402,
-		CR_NEURON_MBUTTON_OUTPUT     = 0x0403
-	} CR_NEURON_TYPE;
-
-	class CRNeuronGroup
+	class CROSContext
 	{
 	protected:
-		static os::CRInputSimulator* s_InputSim;
+		static CROSContext* s_Instance;
+
+		virtual crresult init() = 0;
+
+		virtual void sleep(uint sec, uint ms) = 0;
+		virtual CR_VEC2I getMousePos() = 0;
+		virtual CR_RECT getVirtualScreenClientArea() = 0;
+
+		CROSContext();
 	public:
-		static crresult InitStaticMembers(os::CRInputSimulator* inputSim);
-		static crresult TerminateStaticMembers();
-	protected:
-		uint m_NeuronCount;
+		virtual ~CROSContext();
 
-		CRNeuronGroup(uint neuronCount);
-	public:
-		virtual ~CRNeuronGroup();
+		static crresult InitInstance();
+		static crresult TerminateInstance();
 
-		/*
-		 * Neuron stuff
-		 */
-		/**
-		 * \brief This method processes the input data and returns the result in outData
-		 * 
-		 * \param inData This has to be an array of data with a minimum size of the 
-		 * neuron count. Each element is processed by one neuron.
-		 * \param outData This has to be an array of data with a minimum size of the 
-		 * neuron count. It holds the processed data afterwards.
-		 */
-		virtual void processData(crnwdec const* inData, crnwdec* outData) = 0;
-		/**
-		 * \brief This method is the inverse process from processData. It returns the expected 
-		 * input(outData) for a given output of processData(inData)
-		 * 
-		 * \param inData This has to be an array of data with a minimum size of the 
-		 * neuron count. Each element is processed by one neuron.
-		 * \param outData This has to be an array of data with a minimum size of the 
-		 * neuron count. It holds the expected input data for the output data from
-		 * inData.
-		 */
-		virtual void processDataInverse(crnwdec const* inData, crnwdec* outData) = 0;
-		/**
-		 * \brief This is used in the first creation of the neurons. It should initialize
-		 * the member values with random values.
-		 */
-		virtual void randInit() = 0;
+		inline static void Sleep(uint sec, uint ms = 0)
+		{
+			if (s_Instance) 
+				s_Instance->sleep(sec, ms);
+		}
+		inline static CR_VEC2I GetMousePos()
+		{
+			if (s_Instance)
+				return s_Instance->getMousePos();
 
-		/*
-		 * Getters
-		 */
-		// type stuff and things
-		virtual CR_NEURON_TYPE getType() = 0;
-		bool isType(const CR_NEURON_TYPE& type);
+			return CR_VEC2I(0, 0);
+		}
+		inline static CR_RECT GetVirtualScreenClientArea()
+		{
+			if (s_Instance)
+				return s_Instance->getVirtualScreenClientArea();
+
+			return CR_RECT(0, 0, 0, 0);
+		}
 	};
 
-	typedef cr_ptr<CRNeuronGroup> CRNeuronGroupPtr;
 }}
