+++ CriaAI-Core/src/os/cuda/CudaContext.cuh
